spec helpers extracted to modules
